requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
PyJWT
passlib[bcrypt]